minor tweak
