+++ libs/pandautils/src/main/java/com/instructure/pandautils/utils/Utils.kt
@@ -25,7 +25,6 @@ import android.net.Uri
 import android.widget.Toast
 import com.instructure.canvasapi2.utils.ApiPrefs.domain
 import com.instructure.pandautils.R
-import com.instructure.pandautils.views.CanvasWebView
 import java.io.File
 
 object Utils {
@@ -79,12 +78,5 @@ object Utils {
 
         // Let the user know
         Toast.makeText(context, context.getString(R.string.linkCopied), Toast.LENGTH_SHORT).show()
-    }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    fun loadUrlIntoHeadlessWebView(context: Context, url: String) {
-        val webView = CanvasWebView(context)
-        webView.settings.javaScriptEnabled = true
-        webView.loadUrl(url)
     }
 }
